+++ components/avatar/index.ts
@@ -1,2 +1,3 @@
-export { Avatar, AvatarGroup } from './Avatar'
+export { Avatar } from './Avatar'
+export { AvatarGroup } from './AvatarGroup'
 export type { AvatarProps, AvatarGroupProps, AvatarSize, AvatarShape, AvatarClassNames, AvatarStyles } from './types'